+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.1 under development
 
 - Enh #99: Explicitly import classes and functions in "use" section (@mspirkov)
+- New #100: Add custom SerializerInterface (@Gerych1984)
 
 ## 1.1.0 December 17, 2025